refactor(core): optimize math and saturate operations with platform-specific implementations - Replace manual rounding/clamping with optimized platform-specific implementations for lcvRound, lcvFloor, and lcvCeil - Add lcvClamp helper function and use it in RGB2YCbCr conversion - Remove redundant OMP pragmas and simplify fitLine algorithm - Extend saturate_cast to support int64 and uint64 types - Clean up redundant clip functions in saturate.hpp

+++ liteCV/imgproc/transform.hpp
@@ -21,8 +21,8 @@ namespace lcv
 
         const int width = src.cols;
         const int height = src.rows;
-        int scaled_width = dsize.width ? dsize.width : (int)lcvRound(width * fx);
-        int scaled_height = dsize.height ? dsize.height : (int)lcvRound(height * fy);
+        int scaled_width = dsize.width ? dsize.width : lcvRound(width * fx);
+        int scaled_height = dsize.height ? dsize.height : lcvRound(height * fy);
 
         // scaled width and scaled height must not be zero
         assert(scaled_width * scaled_height != 0);
